Reorganise stack frame layout to avoid unaligned loads.
Ensure 64 bit values are alligned to a 64-bit boundary on the
stack to avoid the unaligned loads penalty.

This is done by where possible moving the first 32 bit value
to the top of the stack when the 64 bit values are unaligned.
If no 32 bit value is available padding is added, this can
cause the stack size to increase however since alignment
padding is added in the stack between the register spill area
and frame entry spilled area often both paddings are not
required so the stack size does not increase.

Changes:
  - Check and add alignment in register allocator

Benchmarks on the Pixel 3 showed gains in VecSumOfMulAdd tests
with a 2.655% and 4.106% reduction in runtime for
VectSumOfMulAdd1 and VectSumOfMulAdd2 respectivly on the silver
core. We also see a 2.386% decrease in the Knucleotide game
benchmark on the sliver core.

Test: benchmarks_run_target.sh --mode 64 --dump-cfg
benchmarks/benchmarksgame/meteor
Test: test_art_target.sh --optimizing

Change-Id: Iadc6689d59053d3a8fb6d54c48321e1a9ad83cba