Add InternOrNewUTF native function
DO NOT SUBMIT. There are ongoing discussions about this mechanism
for exporting ART native functions to frameworks.

This adds a simple mechanism to export native functions from ART to
other platform code, and uses it to provide a facility to
opportunistically avoid allocating duplicate Java strings, when
that can be done without causing delays for weak reference access.

Test: testrunner.py --host -b -t 2045-intern-or-new
Bug: 244750077
Change-Id: Ie3f46eff320666472594e3a0cf65b966ffcbd360
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>